rename firehosenozzle and nozzle
